pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poly-events"
version = "0.1.0"
description = "Polymarket sync workflow for Supabase"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "requests",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
